build.rs: More strictly specify bindgen-exported symbols
`wrapper.h`, while the definitive version of the symbols we want to
generate bindings for, pulls in numerous other symbols that we don't
really care for (for example various `*printf` methods from
`stdio.h`).

An attempt was made to resolve this by blocking `stdlib.h` (now no
longer necessary), but in this case it's likely better to more tightly
control the list of allowed headers via the `allowlist_file`
mechanism instead.

We don't build for openssl compatibility support, so don't expose the
`openssl/` headers. Except for `compat_types.h`, which is required by
other wolfssl bits.