ssl: Wrap `WOLFSSL_CTX` into `WolfSslContext`
Since to create a context you need to pass in a method, wrap the
`*_*_method` bindings too.

I opted to not expose bindings like `wolfDTLS_method` because I'm not
super sure how that would influence later wrappers. It's likely that
my concerns are premature and they can be safely added in later.

Note that these bindings rely on the flags configured within
`build.rs`.